♻️ Refactor accounting list models
Moving pagination details on list models into parent class for less
repetition. Also adding a 'pages()' function to return the pagination
details in a separate object. This will allow us to access pagination
data in the same way across resource types when we implement project
resources, which put the pagination data in a separate place.